# hhspec/specs/architecture/api-contracts.yml
...
openapi: "3.0.3"
info:
  title: "WellAlly Health Web API"
  version: "1.0.0"
  description: "个人医疗健康数据管理 Web 界面后端 API"
servers:
  - url: "http://localhost:3000/api"
    description: "本地开发服务器"
security:
  - bearerAuth: []
paths:
...
  /reports/upload:
    post:
      operationId: uploadReport
      summary: 上传检查单图片并触发 AI 识别
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image, report_date, report_type]
              properties:
                image:
                  type: string
                  format: binary
                  description: 检查单图片 (jpg/png/pdf, ≤20MB)
                report_date:
                  type: string
                  format: date
                report_type:
                  type: string
                  enum: [biochemical, imaging]
                hospital_name:
                  type: string
                  maxLength: 300
                department:
                  type: string
                  maxLength: 200
      responses:
        201:
          description: 上传成功，AI 识别中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthReport"
        400:
          $ref: "#/components/responses/BadRequest"
        413:
          description: 文件过大
...
  /medications:
    get:
      operationId: listMedications
      summary: 查询用药列表
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, discontinued, completed, paused] }
      responses:
        200:
          description: 用药列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medication"
    post:
      operationId: addMedication
      summary: 添加用药记录
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [drug_name, dosage, frequency, start_date]
              properties:
                drug_name: { type: string, maxLength: 300 }
                generic_name: { type: string, maxLength: 300 }
                dosage: { type: string, maxLength: 100 }
                dosage_unit: { type: string, maxLength: 50 }
                frequency: { type: string, maxLength: 100 }
                route: { type: string, enum: [oral, topical, injection, inhalation, sublingual, rectal, other] }
                start_date: { type: string, format: date }
                end_date: { type: string, format: date }
                notes: { type: string }
      responses:
        201:
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medication"
  /medications/{id}:
    put:
      operationId: updateMedication
      summary: 更新用药记录
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        200:
          description: 更新成功
  /medications/interactions/check:
    post:
      operationId: checkInteractions
      summary: 检测当前用药的相互作用
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: 检测结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  interactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/InteractionResult"
                  total_count: { type: integer }
                  x_count: { type: integer, description: "禁忌级数量" }
                  d_count: { type: integer, description: "考虑调整级数量" }

  # ==================== Consultations ====================
...
  /profile:
    get:
      operationId: getProfile
      summary: 获取个人档案
      security: [{ bearerAuth: [] }]
      responses:
        200:
          description: 个人档案
    put:
      operationId: updateProfile
      summary: 更新个人档案
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
                height_cm: { type: number }
                weight_kg: { type: number }
                birth_date: { type: string, format: date }
                gender: { type: string, enum: [male, female, other] }
                blood_type: { type: string, enum: [A, B, AB, O] }
  /profile/allergies:
    get:
      operationId: listAllergies
      summary: 查询过敏史
      security: [{ bearerAuth: [] }]
    post:
      operationId: addAllergy
      summary: 添加过敏记录
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [allergen, severity]
              properties:
                allergen: { type: string, maxLength: 200 }
                severity: { type: string, enum: [mild, moderate, severe, life_threatening] }
                reaction: { type: string }
  /profile/allergies/{id}:
    put:
      operationId: updateAllergy
      security: [{ bearerAuth: [] }]
    delete:
      operationId: deleteAllergy
      security: [{ bearerAuth: [] }]

  # ==================== Settings ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    HealthReport:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        report_type: { type: string }
        title: { type: string }
        report_date: { type: string, format: date }
        hospital_name: { type: string, nullable: true }
        department: { type: string, nullable: true }
        ai_processed: { type: boolean }
        status: { type: string }
        created_at: { type: string, format: date-time }
...
    Medication:
      type: object
      properties:
        id: { type: integer }
        drug_name: { type: string }
        generic_name: { type: string, nullable: true }
        dosage: { type: string }
        dosage_unit: { type: string }
        frequency: { type: string }
        route: { type: string }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date, nullable: true }
        status: { type: string }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    InteractionResult:
      type: object
      properties:
        drug_pair: { type: string }
        severity: { type: string, enum: [A, B, C, D, X] }
        mechanism: { type: string }
        management: { type: string }
        medication_ids: { type: array, items: { type: integer } }
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...